refactor(desktop): split records and tag settings models
